requirements.txt
Django==1.7.4
djangorestframework==3.1.3
Pillow==2.9.0
